requirements.txt
openai
langchain
faiss-cpu
pypdf
tiktoken
langchain-community
langchain_openai
python-docx
# textract